refactor: :recycle: 短縮URLツールをリファクタリング
