+++ src/app/webtool/6/UrlShortener.module.css
@@ -183,10 +183,10 @@
 /* コピー済みアニメーション */
 .copied {
   animation: popPulse 650ms ease forwards;
-  background: rgba(5, 150, 105, 0.18);
-  color: #fff;
-  border-color: rgba(5, 150, 105, 0.28);
-  box-shadow: 0 6px 18px rgba(5, 150, 105, 0.12);
+  background: rgba(5, 150, 105, 0.1);
+  color: var(--primary);
+  border: 1px solid rgba(5, 150, 105, 0.2);
+  box-shadow: 0 4px 12px rgba(5, 150, 105, 0.08);
 }
 
 .copied:hover {
