+++ src/app/webtool/6/types.ts
@@ -1,0 +1,6 @@
+export type ShortenedEntry = {
+  id: string;
+  originalURL: string;
+  shortURL: string;
+  createdAt: string;
+};
